requirements.txt
Flask
Flask-WTF
Flask-Mail
gunicorn
python-linkedin
reportlab
